+++ lib/rubocop/cop/jruby_cops.rb
@@ -1,3 +1,4 @@
 # frozen_string_literal: true
 
 require_relative 'jruby/jar_extension_in_require'
+require_relative 'jruby/snake_case_methods'