+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 1.0.0 (to be released)
 
 * Add new `JRuby/JarExtensionInRequire` cop. ([@Morozzzko][])
+* Add new `JRuby/SnakeCaseMethods` cop. ([@Morozzzko][])